Former-commit-id: 966766ff79f2fe5da06f9df4dcf692528d58394d

+++ another_front/lib/screens/running/widgets/distancebar_custom.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 
 class CustomSliderThumbCircle extends SliderComponentShape {
   final double thumbRadius;
-  final int min;
-  final int max;
+  final double min;
+  final double max;
 
   const CustomSliderThumbCircle({
     required this.thumbRadius,
